DT-326: Push case note creation / amendments to offender event topic (#49)
* DT-326: Push case note creation / amendments to offender event topic

* DT-326: Switch to aws sdk 2 instead

+++ src/main/java/uk/gov/justice/hmpps/casenotes/dto/NomisCaseNote.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Data
 public class NomisCaseNote {
 
-    private Long caseNoteId;
+    private Integer caseNoteId;
     private String offenderIdentifier;
     private String type;
     private String typeDescription;
